Removed circularity of imports by breaking out RPC code into new module
